Move tests inside correct namespace
