+++ tests/Plugin/Prophecy/Token/MaxPriorityTokenTest.php
@@ -1,8 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Tests\Strategy\Prophecy;
-
+namespace Tests\Plugin\Prophecy\Token;
 
 use Moka\Plugin\Prophecy\Token\MaxPriorityToken;
